requirements.txt
spidev
pillow  # Python Imaging Library
numpy

